feat: atualizacao no mapa de locais que contem produto durante a busca
